Changed to a more reusable way of setting up database
Removed the previous versions of setting up my database manually. Added in forms script to hold all the forms to be added. Created database using SQLAlchemy with db_setup script for the initial database setup. Added the products form and ability to add new products to the database using the form