Implementation of delete and edit for users page
A delete and edit button have been added beside each record in user table. Upon clicking delete a pop up is displayed to confirm, if user confirms the row is deleted from database. The edit button adds the selected user details to the form, these can then be edited, and will be updated in the database upon pressing submit.