Delete and edit functions added to all relevant pages
Delete button added to sales tracker and database. Edit button added to database and sales tracker, with form being filled in with relevant data to edit.

Have not added this functionality to targets page as updating store targets will be based on current user login, and individual targets will be added and updated based on hours added