Implemented update progress on update/delete sale
Added to the update progress method to alter progress for users/store when a sale is deleted or updated. The method now takes an extra parameter to for either delete or adding progress from a sale. On editing the sale it is first deleted with the original data, then added with the updated sale data.