User authentication implemented
Users inputted username and password are verified against the stored username and password hash. Admin and sales users are redirected to specified pages after sign in. Users are only able to access any other pages of the application once the session is authenticated. Users session is deleted and redirected to login page with logout button or after 1 hour.