Commission feature updated
Commission is now based on the attributes of the individual sale. A commission column has been added to the progress table to store individual users total commission, and to the sales table to track the commission calculation. This calculation adds to the commission value within the products table, and is updated or deleted as the sale is.