Bug resolved. Deleted users do not break sales tracker
A new column has been added to users and products table called is_deleted. This is set to True when user/product is deleted instead of removing from database. Users access is revoked when they are deleted. Users and products are still accessible in table where needed, but are removed from selection if is_deleted is set to true