revenue multiplier implemented
Revenue column added to sales table to make calculation based on sales attributes and product revenue. calc_rev method developed which is called each time a sale or product is updated. This is then displayed on the sales tracker page and added to the revenue progress to be displayed on the targets page